Clarify config settings for AS::halt_callback_chains_on_return_false
and wrap lines at 80 characters.

Despite the config name, it isn't super clear when to use true or false
for this config and which the two is the new default... particularly
with the word 'false' in the config name.

[skip ci]

Move docs from the initializer to the Configuring Guide [skip ci]

Note that the Configuring Guide previously indicated that
ActiveSupport.halt_callback_chains_on_return_false defaulted to true.
This PR changes it to state that the default setting is false.